+++ android/app/src/main/AndroidManifest.xml
@@ -16,7 +16,7 @@
         android:usesCleartextTraffic="true"
         tools:targetApi="31">
         <activity
-            android:name=".presentation.studyList.StudyDetail"
+            android:name=".presentation.studyDetail.StudyDetailActivity"
             android:exported="false" />
         <activity
             android:name=".presentation.MainActivity"
@@ -30,7 +30,7 @@
 
         <activity
             android:name=".presentation.createStudy.CreateStudyActivity"
-            android:exported="true" />
+            android:exported="false" />
     </application>
 
 </manifest>
